+++ src/lib/presets/vanilla.js
@@ -1,0 +1,14 @@
+import defaultStyles from '../default.css';
+import List from '../List.svelte';
+import Item from '../Item.svelte';
+import Selection from '../Selection.svelte';
+import Multi from '../Multi.svelte';
+
+
+export default {
+    theme: defaultStyles,
+    List,
+    Item,
+    Selection,
+    Multi
+}