+++ test/src/Select/ParentContainer.svelte
@@ -3,10 +3,11 @@
 
   export let value;
   export let items;
+  export let config;
 </script>
 
 <div class="container">
-  <Select bind:value {items}></Select>
+  <Select bind:value {items} {config}></Select>
 
   <p class="result">{value.label}</p>
 </div>